+++ render/src/plane.rs
@@ -1,8 +1,8 @@
 use nalgebra::Vector3;
 // for defining the frustum bounding planes
 pub struct Plane {
-    pub normal: Vector3<f64>,
-    pub distance: f64,
+    pub normal: Vector3<f32>,
+    pub distance: f32,
 }
 
 
@@ -17,7 +17,7 @@ RIGHT HAND coord system
 
 impl Plane {
     //https://www.lighthouse3d.com/tutorials/maths/plane/
-    pub fn from_points(p0:Vector3<f64>, p1: Vector3<f64>, p2: Vector3<f64>) -> Plane{
+    pub fn from_points(p0:Vector3<f32>, p1: Vector3<f32>, p2: Vector3<f32>) -> Plane{
         let v = p1 - p0;
         let u = p2 - p0;
         let n = v.cross(&u).normalize();
@@ -30,7 +30,7 @@ impl Plane {
         }
     }
 
-    pub fn is_inside(&self, p: Vector3<f64> ) -> bool{
+    pub fn is_inside(&self, p: Vector3<f32> ) -> bool{
 
         self.normal.dot(&p) - self.distance <=0.0
     }
